refactor: 변경된 설계 구조에 맞게 클래스 패키지 이동
