+++ src/main/java/com/example/demo/dto/BusinessHourResponse.java
@@ -1,14 +1,14 @@
-package com.example.demo.service.dto;
+package com.example.demo.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class BusinessHourDto {
+public class BusinessHourResponse {
 
 	private final String day;
 	private final String startTime;
 	private final String endTime;
-	
+
 }
