+++ src/main/java/com/example/demo/dto/CafeSearchCondition.java
@@ -1,4 +1,4 @@
-package com.example.demo.service.dto;
+package com.example.demo.dto;
 
 import com.example.demo.domain.MaxAllowableStay;
 import com.example.demo.domain.MinMenuPrice;
