+++ src/main/java/com/example/demo/dto/SnsResponse.java
@@ -1,11 +1,11 @@
-package com.example.demo.service.dto;
+package com.example.demo.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class SnsDto {
+public class SnsResponse {
 
 	private final String name;
 	private final String url;
