+++ src/main/java/com/example/demo/service/CafeQueryService.java
@@ -1,11 +1,11 @@
 package com.example.demo.service;
 
+import com.example.demo.dto.CafeSearchRequest;
+import com.example.demo.dto.CafeSearchResponse;
 import com.example.demo.dto.PagedResponse;
-import com.example.demo.service.dto.CafeSearchResponse;
-import com.example.demo.service.dto.ServiceCafeSearchRequest;
 
 public interface CafeQueryService {
 
-	public PagedResponse<CafeSearchResponse> searchWithPagingByDynamicFilter(ServiceCafeSearchRequest request);
+	public PagedResponse<CafeSearchResponse> searchWithPagingByDynamicFilter(CafeSearchRequest request);
 
 }
